requirements.txt
asgiref
crispy-bootstrap5
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
Markdown
MouseInfo
mss
numpy
opencv-contrib-python
Pillow
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
python-decouple
pytweening
pytz
sqlparse
typing_extensions
tzdata